requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.11.0
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
